updated after mySQL crash
Fully functional pages minus deducting from database. Will continue working on matching CSS, but wanted to get the fully operational pages up here. We will need to get this merch database on Sydney's laptop and test connecting our pages at some point this week. The computer we use to present (Sydney's) has to be the one with the database and merch table on it. My pages only work if the database is on the local machine.